+++ js-library/src/model/environments.enum.ts
@@ -1,4 +1,5 @@
 export enum Environments {
   LOCALHOST,
   GOERLI,
+  SEPOLIA,
 }