+++ js-library/src/services/dapp-registry.ts
@@ -1,6 +1,7 @@
 import { utils, Contract, providers, BigNumber } from 'ethers'
 import { DAPP_REGISTRY_ENVIRONMENT_CONFIGS } from '../constants'
 import DappRegistryContractLocal from '../contracts/DappRegistry/DappRegistry.json'
+import RatingsContractLocal from '../contracts/Ratings/Ratings.json'
 import {
   DappRegistryEnvironment,
   Environments,
@@ -12,8 +13,10 @@ import {
 import { DappRecord } from '../model/dapp-record.model'
 import { waitTransaction } from '../utils/tx'
 import { SignerOrProvider } from './ens'
+import { Rating } from '../model/rating.model'
 
 export const DappRegistryContract = DappRegistryContractLocal
+export const RatingsContract = RatingsContractLocal
 
 const ADMIN_ROLE = utils.hexZeroPad(utils.hexlify(0), 32)
 const VALIDATOR_ROLE = utils.keccak256(utils.toUtf8Bytes('VALIDATOR_ROLE'))
@@ -25,6 +28,7 @@ const VALIDATOR_ROLE = utils.keccak256(utils.toUtf8Bytes('VALIDATOR_ROLE'))
 export class DappRegistry {
   private _provider: providers.JsonRpcProvider
   private _dappRegistryContract: Contract
+  private _ratingsContract: Contract
 
   constructor(
     config: DappRegistryEnvironment = DAPP_REGISTRY_ENVIRONMENT_CONFIGS[Environments.LOCALHOST],
@@ -33,10 +37,12 @@ export class DappRegistry {
     this._provider = new providers.JsonRpcProvider(config.rpcUrl)
 
     this._dappRegistryContract = new Contract(
-      config.contractAddress,
+      config.dappRegistryAddress,
       DappRegistryContract.abi,
       this._provider,
     )
+
+    this._ratingsContract = new Contract(config.ratingsAddress, RatingsContract.abi, this._provider)
 
     if (signerOrProvider) {
       this.connect(signerOrProvider)
@@ -56,6 +62,7 @@ export class DappRegistry {
    */
   public connect(signerOrProvider: SignerOrProvider): void {
     this._dappRegistryContract = this._dappRegistryContract.connect(signerOrProvider)
+    this._ratingsContract = this._ratingsContract.connect(signerOrProvider)
   }
 
   public grantAdminRole(address: EthAddress): Promise<void> {
@@ -126,6 +133,30 @@ export class DappRegistry {
     return this._dappRegistryContract.getUser(address)
   }
 
+  public rateDapp(recordLocation: SwarmLocation, review: HexString, rating: number): Promise<void> {
+    return waitTransaction(this._ratingsContract.rate(recordLocation, review, rating))
+  }
+
+  public async getAverageRating(recordLocation: SwarmLocation): Promise<number> {
+    const averageRating = Number(await this._ratingsContract.getAverageRating(recordLocation))
+
+    return Number.isInteger(averageRating) ? averageRating : 0
+  }
+
+  public async getNumberOfRatings(recordLocation: SwarmLocation): Promise<BigNumber> {
+    return this._ratingsContract.getNumberOfRatings(recordLocation)
+  }
+
+  public async hasUserRated(recordLocation: SwarmLocation): Promise<boolean> {
+    return Boolean(await this._ratingsContract.hasUserRated(recordLocation))
+  }
+
+  public async getRatingFor(recordLocation: SwarmLocation): Promise<Rating[]> {
+    const records = await this._ratingsContract.getRatingFor(recordLocation)
+
+    return records.map((record: string[]) => this.convertRatingRecord(record))
+  }
+
   private convertDappRecord(record: Array<string>): DappRecord {
     return {
       recordHash: record[0],
@@ -136,6 +167,14 @@ export class DappRegistry {
       index: BigNumber.from(record[5]),
       creatorIndex: BigNumber.from(record[6]),
       timestamp: new Date(record[7]),
+    }
+  }
+
+  private convertRatingRecord(record: Array<string>): Rating {
+    return {
+      rating: Number(record[0]),
+      review: record[1],
+      user: record[2],
     }
   }
 }