+++ scripts/save-addresses.sh
@@ -22,6 +22,7 @@ FDS_REGISTRAR_ADDRESS=$(echo "$2" | grep -Po 'FDSRegistrar deployed to: \K[^\s]*
 ENS_REGISTRY_ADDRESS=$(echo "$2" | grep -Po 'ENSRegistry deployed to: \K[^\s]*')
 PUBLIC_RESOLVER_ADDRESS=$(echo "$2" | grep -Po 'PublicResolver deployed to: \K[^\s]*')
 DAPP_REGISTRY_ADDRESS=$(echo "$2" | grep -Po 'DappRegistry deployed to: \K[^\s]*')
+RATINGS_ADDRESS=$(echo "$2" | grep -Po 'Ratings deployed to: \K[^\s]*')
 
 if [[ -z "$FDS_REGISTRAR_ADDRESS" || -z "$ENS_REGISTRY_ADDRESS" || -z "$PUBLIC_RESOLVER_ADDRESS" ]];
 then
@@ -37,10 +38,17 @@ then
   DAPP_REGISTRY_ADDRESS=$(echo "$EXISTING_ADDRESSES" | grep -Po '_DAPP_REGISTRY_ADDRESS=\K[^\s]*')
 fi
 
+if [[ -z "$RATINGS_ADDRESS" ]];
+then
+  # Extracting contract addresses from existing file
+  RATINGS_ADDRESS=$(echo "$EXISTING_ADDRESSES" | grep -Po '_RATINGS_ADDRESS=\K[^\s]*')
+fi
+
 # Saving contract addresses to an .env file
 echo "${1^^}_ENS_REGISTRY_ADDRESS=$ENS_REGISTRY_ADDRESS" >> "$ENV_FILE"
 echo "${1^^}_FDS_REGISTRAR_ADDRESS=$FDS_REGISTRAR_ADDRESS" >> "$ENV_FILE"
 echo "${1^^}_PUBLIC_RESOLVER_ADDRESS=$PUBLIC_RESOLVER_ADDRESS" >> "$ENV_FILE"
 echo "${1^^}_DAPP_REGISTRY_ADDRESS=$DAPP_REGISTRY_ADDRESS" >> "$ENV_FILE"
+echo "${1^^}_RATINGS_ADDRESS=$RATINGS_ADDRESS" >> "$ENV_FILE"
 sed -i '/^$/d' "$ENV_FILE"
 echo "Contract addresses saved to: $ENV_FILE"